% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.hisat2Bin}
\alias{.hisat2Bin}
\title{Call the HISAT2 binary with additional arguments.}
\usage{
.hisat2Bin(bin = c("hisat2", "hisat2-build"), args = "",
  execute = TRUE)
}
\arguments{
\item{bin}{The name of the binary, either hisat2 or hisat2-build}

\item{args}{A character string containing the arguments that will be passed
to the binary}

\item{execute}{Logical scalar, whether to execute the command. If FALSE,
return a string with the shell command.}
}
\value{
If \code{execute} is TRUE, returns the console output of running the
  hisat2 command. If \code{execute} is FALSE, returns the shell command.
}
\description{
Adapted from the Rbowtie package
}
\keyword{internal}
