% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{hisat2}
\alias{hisat2}
\title{Align reads with HISAT2}
\usage{
hisat2(sequences, index, ..., type = c("single", "paired", "crossbow"),
  outfile, force = FALSE, strict = TRUE)
}
\arguments{
\item{sequences}{Character vector (for single-end experiments) or list of two
character vectors (for paired-end experiments) of file names of fasta or
fastq files that will be aligned by HISAT2. The character vectors can contain
more than one file name, in which case the reads will be combined across
all the provided files in the corresponding vector.}

\item{index}{Character scalar, giving the directory where the genome
index is stored.}

\item{...}{Additional arguments provided to STAR in the genome index
generation.}

\item{type}{Character scalar, either "single" or "paired", giving the type of
the experiment.}

\item{outfile}{(optional) Character scalar, the prefix of the
alignment files (including the output directory). If not provided, will be
set to the current directory (\code{./}).}

\item{force}{Logical scalar, whether or not to overwrite an existing output
directory.}

\item{strict}{Logical scalar, whether strict checking of input arguments
should be enforced.}

\item{execute}{Logical scalar, whether to execute the command. If FALSE,
return a string with the shell command.}
}
\value{
If \code{execute} is TRUE, invisibly returns the console output of
  running the read alignment. If \code{execute} is FALSE, invisibly returns
  the shell command.
}
\description{
Align reads with HISAT2
}
\examples{
\dontrun{
starAlign(readFilesIn="", genomeDir="genome", type="single",
          outFileNamePrefix="output/sample1")
}
}
\author{
Charlotte Soneson
}
