% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{hisat2_build}
\alias{hisat2_build}
\title{Generate HISAT2 genome index}
\usage{
hisat2_build(references, outdir, ..., prefix = "index", force = FALSE,
  strict = TRUE, execute = TRUE)
}
\arguments{
\item{references}{Character vector of file names of fasta files that will be
indexed by HISAT2.}

\item{outdir}{Character scalar, giving the directory where the genome
index will be stored.}

\item{...}{Additional arguments provided to STAR in the genome index
generation.}

\item{force}{Logical scalar, whether or not to overwrite an existing index
directory.}

\item{strict}{Logical scalar. Whether strict checking of input arguments
should be enforced.}

\item{execute}{Logical scalar, whether to execute the command. If FALSE,
return a string with the shell command.}
}
\value{
If \code{execute} is TRUE, invisibly returns the console output of
  running the genome indexing. If \code{execute} is FALSE, invisibly returns
  the shell command.
}
\description{
Generate HISAT2 genome index
}
\examples{
\dontrun{
hisat2_build(fastaFiles="genome.fa", genomeDir="genome")
}
}
\author{
Charlotte Soneson
}
