% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{hisat2_build_usage}
\alias{hisat2_build_usage}
\title{Print usage of hisat2-build}
\usage{
hisat2_build_usage()
}
\description{
Print usage of hisat2-build
}
\examples{
hisat2_build_usage()
}
\author{
Charlotte Soneson
}
