% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{hisat2_usage}
\alias{hisat2_usage}
\title{Print usage of hisat2}
\usage{
hisat2_usage()
}
\value{
No value is returned, the usage of \code{hisat2} is printed to the
  console.
}
\description{
Print usage of hisat2
}
\examples{
hisat2_usage()
}
\author{
Charlotte Soneson
}
