% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2.R
\name{hisat2_version}
\alias{hisat2_version}
\title{Print HISAT2 version}
\usage{
hisat2_version()
}
\value{
No value is returned, the version information for \code{hisat2} is
  printed to the console.
}
\description{
Print HISAT2 version
}
\examples{
hisat2_version()
}
\author{
Charlotte Soneson
}
