% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhisat2-package.R
\docType{package}
\name{Rhisat2-package}
\alias{Rhisat2}
\alias{Rhisat2-package}
\title{R hisat2 wrapper}
\description{
This package provides an R wrapper around the HISAT2 short read aligner. The
package is used by the QuasR Bioconductor package. We recommend to use QuasR
instead of calling Rhisat2 directly, since it covers the whole analysis
workflow of typical high-throughput sequencing experiments.
}
\references{
Kim D, Langmead B and Salzberg SL. HISAT: a fast spliced aligner with low
memory requirements. Nature Methods 12:357-360 (2015).
}
\seealso{
hisat2, hisat2_build, hisat2_build_usage, hisat2_usage
}
\author{
Charlotte Soneson, based partly on code from Florian Hahne
}
\keyword{internal}
